endpointmanager: Implement subscriber.Node for host endpoint
This commit converts the EndpointManager into a subscriber for node events
from K8s, which handles specific host endpoint logic. With this change,
the logic that was previously inside the K8s node watcher has now been
copied here.

Future commits will augment the K8s node watcher to notify subscribers
to node events, thereby transferring the logic from the K8s watcher to
this package.
